xml/parsing/manakai/condsects-3.dat

#data
<!DOCTYPE a [
  <![IGNORE[
    <!ENTITY x "X">
  ]]>
  <!ENTITY y "Y">
]><a>&x;&y;</a>
#errors
1;1;no XML decl;;;s
2;3;condsect in internal subset;;;m
6;6;entity not declared;;x;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a [
  <![INCLUDE[
    <!ENTITY x "X">
  ]]>
  <!ENTITY y "Y">
]><a>&x;&y;</a>
#errors
1;1;no XML decl;;;s
2;3;condsect in internal subset;;;m
6;6;entity not declared;;x;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a [
  <!ENTITY % hoge '
    <![INCLUDE[
      <!ENTITY x "X">
    ]]>
    <!ENTITY y "Y">
  '>
  %hoge;
]><a>&x;&y;</a>
#errors
1;1;no XML decl;;;s
3;5;condsect in internal subset;;;m
6;5;xml:dtd:ext decl;;;w
9;6;entity not declared;;x;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "&x;Y"

#data
<!DOCTYPE a [
  <!ENTITY % hoge SYSTEM "http://b/">
  %hoge;
]><a>&x;&y;</a>
#resource http://b/
    <![INCLUDE[
      <!ENTITY x "X">
    ]]>
    <!ENTITY y "Y">
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;7;xml:dtd:ext decl;;;w
[1]4;5;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a>
| <{} a>
|   "XY"

#data
<!DOCTYPE a [
  <!ENTITY % fuga '
    <![INCLUDE[
      <!ENTITY x "X">
    ]]>
    <!ENTITY y "Y">
  '>
  <!ENTITY % hoge SYSTEM "http://b/">
  %hoge;
]><a>&x;&y;</a>
#resource http://b/
%fuga;
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
4;7;xml:dtd:ext decl;;;w
6;5;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a>
| <{} a>
|   "XY"

#data
<!DOCTYPE a [
  <![IGNORE[
    <!ENTITY x "X">
]><a/>
#errors
1;1;no XML decl;;;s
2;3;condsect in internal subset;;;m
4;1;ms:bare ];;;m
4;7;ms:unclosed;;;m
4;7;unclosed internal subset;;;m
4;7;no root element;;;m
#document
| <!DOCTYPE a>
#entities

#data
<!DOCTYPE a [
  <![INCLUDE[
    <!ENTITY x "X">
]><a/>
#errors
1;1;no XML decl;;;s
2;3;condsect in internal subset;;;m
4;1;ms:bare ];;;m
4;7;ms:unclosed;;;m
4;7;unclosed internal subset;;;m
4;7;no root element;;;m
#document
| <!DOCTYPE a>
#entities
